move locales-dropdown component to common folder
